% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animalesSmall}
\alias{animalesSmall}
\title{Small animals dataset (spanish).}
\description{
Small spanish dataset of animals.
}
\author{
Rapha\"{e}l Couturier \email{raphael.couturier@univ-fcomte.fr}
}

