% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callCFunctions.R
\name{callAsirules}
\alias{callAsirules}
\title{Calls the C++ ASI rules processor.}
\usage{
callAsirules()
}
\description{
Interface to call the ASI rules processor.
}
\author{
Raphael Couturier
}

