\name{callHierarchyComputation}
\alias{callHierarchyComputation}
\title{Calls the C++ hierarchy computation.}
\usage{
callHierarchyComputation(cohesion_matrix, list.selected.item,
  list.occurrences.variables)
}
\description{
Interface to call the the C++ hierarchy computation.
}
\author{
Rapha\"{e}l Couturier
\email{raphael.couturier@univ-fcomte.fr}
}

