% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data2transac}
\alias{data2transac}
\title{Saves the data to a transaction file.}
\usage{
data2transac(data)
}
\arguments{
\item{data}{Matrix of data.}
}
\description{
Saves the data to a transaction file.
}
\author{
Rapha\"{e}l Couturier \email{raphael.couturier@univ-fcomte.fr}
}

