% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hierarchyTree}
\alias{hierarchyTree}
\title{Computes and Displays the Hierarchy Tree.}
\usage{
hierarchyTree(list.variables, Supplementary.variables, Matrix.values,
  Verbose = FALSE)
}
\arguments{
\item{list.variables}{list of variables to compute the similarity tree from.}

\item{Supplementary.variables}{list of supplementary variables.}

\item{Matrix.values}{matrix with values of individuals (used to compute the contributions and typicalities).}

\item{Verbose}{give many details.}
}
\description{
Reads the ASI rules, computes the hierarchy tree and displays it.
}
\author{
Rapha\"{e}l Couturier \email{raphael.couturier@univ-fcomte.fr}
}

