% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/implicativeGraph.R
\name{implicativeGraph}
\alias{implicativeGraph}
\title{Displays the Implicative Graph.}
\usage{
implicativeGraph(list.variables, computing.mode = 1)
}
\arguments{
\item{list.variables}{list containing all the variables used in the computation}

\item{computing.mode}{controls the computing mode: 1=classic implication, 2=classic implication+ confidence, 3=implifiance}
}
\description{
Reads the ASI rules, selects the rules according to the toolbar and calls rgraphviz before displaying the rules.
}
\author{
Raphaël Couturier \email{raphael.couturier@univ-fcomte.fr}
}

