% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCFunctions.R
\name{callAsirules}
\alias{callAsirules}
\title{Calls the C++ write_transactions.}
\usage{
callAsirules()
}
\arguments{
\item{x}{dataframe representing all the data}
}
\description{
Interface to call the the C++ write_transactions.
}
\author{
Raphael Couturier
}
