% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCFunctions.R
\name{callHierarchyComputation}
\alias{callHierarchyComputation}
\title{Calls the C++ hierarchy computation.}
\usage{
callHierarchyComputation(
  cohesion_matrix,
  list.occurrences.variables,
  supplementary.variables,
  matrix.values,
  contribution.supp,
  typicality.supp,
  verbose
)
}
\arguments{
\item{cohesion_matrix}{cohesion matrix of the variables}

\item{list.occurrences.variables}{list of the occurrences of the variables.}

\item{supplementary.variables}{list of supplementary variables.}

\item{matrix.values}{matrix with values of individuals (used to compute the contributions and typicalities).}

\item{contribution.supp}{boolean to compute the contribution of supplementary variables}

\item{typicality.supp}{boolean to compute the typicality of supplementary variables}

\item{verbose}{gives more details}
}
\description{
Interface to call the the C++ hierarchy computation.
}
\author{
Raphael Couturier
}
