% Generated by roxygen2 (4.0.1): do not edit by hand
\name{callSimilarityComputation}
\alias{callSimilarityComputation}
\title{Calls the C++ similarity computation.}
\usage{
callSimilarityComputation(similarity_matrix, list.occurrences.variables,
  supplementary.variables, matrix.values, verbose)
}
\arguments{
\item{similarity_matrix}{matrix of similarities of the variables.}

\item{list.occurrences.variables}{list of the occurrences of the variables.}

\item{verbose}{gives more details}
}
\description{
Interface to call the the C++ similarity computation.
}
\author{
Rapha\"{e}l Couturier \email{raphael.couturier@univ-fcomte.fr}
}

