\name{callSimilarityComputation}
\alias{callSimilarityComputation}
\title{Calls the C++ similarity computation.}
\usage{
callSimilarityComputation(similarity_matrix, list.selected.item,
  list.occurrences.variables)
}
\description{
Interface to call the the C++ similarity computation.
}
\author{
Rapha\"{e}l Couturier
\email{raphael.couturier@univ-fcomte.fr}
}

