% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCFunctions.R
\name{callSimilarityComputation}
\alias{callSimilarityComputation}
\title{Calls the C++ ASI rules processor.}
\usage{
callSimilarityComputation(
  similarity_matrix,
  list.occurrences.variables,
  supplementary.variables,
  matrix.values,
  contribution.supp,
  typicality.supp,
  verbose
)
}
\arguments{
\item{similarity_matrix}{matrix of similarities of the variables.}

\item{list.occurrences.variables}{list of the occurrences of the variables.}

\item{supplementary.variables}{list of supplementary variables.}

\item{matrix.values}{matrix with values of individuals (used to compute the contributions and typicalities).}

\item{contribution.supp}{boolean to compute the contribution of supplementary variables}

\item{typicality.supp}{boolean to compute the typicality of supplementary variables}

\item{verbose}{gives more details}
}
\description{
Interface to call the ASI rules processor.

Interface to call the the C++ similarity computation.
}
\author{
Raphael Couturier
}
