% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2transac.R
\name{data2transac}
\alias{data2transac}
\title{Saves the data to a transaction file.}
\usage{
data2transac(data)
}
\arguments{
\item{data}{Matrix of data.}
}
\description{
Saves the data to a transaction file.
}
\author{
Raphael Couturier
}
