% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hierarchyTree}
\alias{hierarchyTree}
\title{Computes and Displays the Hierarchy Tree.}
\usage{
hierarchyTree(list.variables, Verbose = FALSE)
}
\arguments{
\item{list.variables}{list of variables to compute the hierarchy tree from.}
}
\description{
Reads the ASI rules, computes the hierarchy tree and displays it.
}
\author{
Rapha\"{e}l Couturier \email{raphael.couturier@univ-fcomte.fr}
}

