% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyTree.R
\name{hierarchyTree}
\alias{hierarchyTree}
\title{Computes and Displays the Hierarchy Tree.}
\usage{
hierarchyTree(
  fileName,
  list.variables,
  supplementary.variables,
  matrix.values,
  contribution.supp,
  typicality.supp,
  computing.mode = 1,
  verbose = FALSE
)
}
\arguments{
\item{fileName}{name of the file containing data}

\item{list.variables}{list of variables to compute the similarity tree from.}

\item{supplementary.variables}{list of supplementary variables.}

\item{matrix.values}{matrix with values of individuals (used to compute the contributions and typicalities).}

\item{contribution.supp}{boolean to compute the contribution of supplementary variables}

\item{typicality.supp}{boolean to compute the typicality of supplementary variables}

\item{computing.mode}{controls the computing mode: 1=classic implication, 2=classic implication+ confidence, 3=implifiance}

\item{verbose}{give many details.}
}
\description{
Reads the ASI rules, computes the hierarchy tree and displays it.
}
\author{
Raphael Couturier
}
