% Generated by roxygen2 (4.0.1): do not edit by hand
\name{implicativeGraph}
\alias{implicativeGraph}
\title{Displays the Implicative Graph.}
\usage{
implicativeGraph(list.variables)
}
\arguments{
\item{list.variables}{list containing all the variables used in the computation}
}
\description{
Reads the ASI rules, selects the rules according to the toolbar and calls rgraphviz before displaying the rules.
}
\author{
Rapha\"{e}l Couturier \email{raphael.couturier@univ-fcomte.fr}
}

