\name{implicativeGraph}
\alias{implicativeGraph}
\title{Displays the Implicative Graph.}
\usage{
implicativeGraph(list.variables)
}
\arguments{
  \item{rules}{dataframe of ASI rules.}
}
\description{
Reads the ASI rules, selects the rules according to the
toolbar and calls rgraphviz before displaying the rules.
}
\author{
Rapha\"{e}l Couturier
\email{raphael.couturier@univ-fcomte.fr}
}

