\name{similarityTree}
\alias{similarityTree}
\title{Computes and Displays the Similarity Tree.}
\usage{
similarityTree(rules = NULL)
}
\arguments{
  \item{rules}{dataframe of ASI rules.}
}
\description{
(Reads the ASI rules) Computes the similarities and
displays the Similarity Tree.
}
\author{
Rapha\"{e}l Couturier
\email{raphael.couturier@univ-fcomte.fr}
}

