\name{similarityTree}
\alias{similarityTree}
\title{Computes and Displays the Similarity Tree.}
\usage{
similarityTree(list.variables, rules = NULL, Verbose = FALSE)
}
\arguments{
  \item{list.variables}{list of variables to compute the
  similarity tree from}

  \item{rules}{dataframe of ASI rules.}

  \item{verbose}{give many details}
}
\description{
(Reads the ASI rules) Computes the similarities and
displays the Similarity Tree.
}
\author{
Rapha\"{e}l Couturier
\email{raphael.couturier@univ-fcomte.fr}
}

