% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityTree.R
\name{similarityTree}
\alias{similarityTree}
\title{Computes and Displays the Similarity Tree.}
\usage{
similarityTree(
  fileName,
  list.variables,
  supplementary.variables,
  matrix.values,
  contribution.supp,
  typicality.supp,
  verbose = FALSE
)
}
\arguments{
\item{fileName}{name of the file containing data}

\item{list.variables}{list of variables to compute the similarity tree from.}

\item{supplementary.variables}{list of supplementary variables.}

\item{matrix.values}{matrix with values of individuals (used to compute the contributions and typicalities).}

\item{contribution.supp}{boolean to compute the contribution of supplementary variables}

\item{typicality.supp}{boolean to compute the typicality of supplementary variables}

\item{verbose}{give many details.}
}
\description{
(Reads the ASI rules) Computes the similarities and displays the Similarity Tree.
}
\author{
Raphael Couturier
}
